+++ cosmohub/api/marshal/__init__.py
@@ -1,4 +1,0 @@
-from . import (
-    fields,
-    schema,
-)