+++ cosmohub/api/rest/catalogs.py
@@ -6,13 +6,14 @@ from sqlalchemy.orm import joinedload
 
 from cosmohub.api import db, api_rest
 
-from ..marshal import schema
+from .downloads import DatasetReadmeDownload, FileReadmeDownload, FileContentsDownload
+from .. import fields
 from ..db import model
 from ..db.session import transactional_session
-from ..security import auth_required, PRIV_USER
+from ..security import auth_required, Privilege, Token
 
 class CatalogCollection(Resource):
-    decorators = [auth_required(PRIV_USER)]
+    decorators = [auth_required(Privilege(['user']))]
 
     def get(self):
         with transactional_session(db.session, read_only=True) as session:
@@ -23,17 +24,17 @@ class CatalogCollection(Resource):
             restricted = session.query(model.Catalog).join(
                 'groups', 'users'
             ).filter(
-                model.User.id == getattr(g, 'current_user')['id']
+                model.User.id == g.session['user'].id
             )
 
             catalogs = public.union(restricted).all()
 
-            return marshal(catalogs, schema.CatalogCollection)
+            return marshal(catalogs, fields.CatalogCollection)
 
 api_rest.add_resource(CatalogCollection, '/catalogs')
 
 class CatalogItem(Resource):
-    decorators = [auth_required(PRIV_USER)]
+    decorators = [auth_required(Privilege(['user']))]
 
     def get(self, id_):
         with transactional_session(db.session, read_only=True) as session:
@@ -49,16 +50,38 @@ class CatalogItem(Resource):
                     'groups', 'catalogs'
                 ).filter(
                     model.Catalog.id == id_,
-                    model.User.id == getattr(g, 'current_user')['id'],
+                    model.User.id == g.session['user'].id,
                 ).first()
 
                 if not user:
                     raise http_exc.Forbidden
 
             columns = current_app.columns.loc[catalog.relation].to_dict('records')
-            data = marshal(catalog, schema.Catalog)
+            data = marshal(catalog, fields.Catalog)
             data.update({'columns' : columns})
-
+            
+            for dataset in data['datasets']:
+                token = Token(
+                    g.session['user'],
+                    Privilege(['download'], ['dataset'], [dataset['id']]),
+                    expires_in=current_app.config['TOKEN_EXPIRES_IN']['download'],
+                )
+                    
+                url = api_rest.url_for(DatasetReadmeDownload, id_=dataset['id'], auth_token=token.dump(), _external=True)
+                dataset['download_readme'] = url
+            
+            for file_ in data['files']:
+                token = Token(
+                    g.session['user'],
+                    Privilege(['download'], ['file'], [file_['id']]),
+                    expires_in=current_app.config['TOKEN_EXPIRES_IN']['download'],
+                )
+                
+                url = api_rest.url_for(FileReadmeDownload, id_=file_['id'], auth_token=token.dump(), _external=True)
+                file_['download_readme'] = url
+                url = api_rest.url_for(FileContentsDownload, id_=file_['id'], auth_token=token.dump(), _external=True)
+                file_['download_contents'] = url
+            
             return data
 
 api_rest.add_resource(CatalogItem, '/catalogs/<int:id_>')